Fixed an error where  intents are required.
In newer versions (2023) when creating a bot in order to send a message to the server, user is required to specify which intents to give him for bot to be able to send messages, menage channels, etc. With this fix, after going into the discord developer portal and granting the bot message content intent, bot will be able to send messages to the server with this fix.